fix: Service 변경에 따라 연관된 코드의 반환 타입을 long으로 수정

+++ src/main/java/calculator/Controller/CalculatorController.java
@@ -10,7 +10,7 @@ public class CalculatorController {
             String input = InputView.readInput();
 
             CalculatorService calculator = new CalculatorService();
-            int result = calculator.calculate(input);
+            Long result = calculator.calculate(input);
 
             OutputView.printResult(result);
